drupal 9.2.7 -> 9.2.10
(Obtained from : https://www.drupal.org/project/drupal/releases/8.x-9.2.10) 

This is a **patch (bugfix) release of Drupal 9** and is ready for use on production sites. [Learn more about Drupal 9].

Drupal 9.2.x will receive security coverage until June 2022 when Drupal 9.4.0 is released.

If you are upgrading from Drupal 8, read [upgrading a Drupal 8 site to Drupal 9], [9.0.0 release notes], and the [9.2.0 release notes] before upgrading to this release.

Regardless of which version you choose now, features will only be added to Drupal 9 minor releases, so plan to adopt Drupal 9 as soon as possible so that you can easily update to Drupal 9.2 and later.

This release updates Symfony components to [4.4.35] (where applicable). Note that the Symfony CVE fixed by 4.4.35 does not affect Drupal core, so this is being released as a regular patch release.

### Known issues

[Search the issue queue for known issues].